# com-bookstore-task/src/main/java/combookstoretask/request/BookRequest.java
package combookstoretask.request;

import lombok.Data;

@Data
public class BookRequest {
	
	private String bookName;
	
	private String BookCategory;
	
	private String bookAuthor;
	
	private String bookDetails;
	
	
	
	

}